refactor(relay): 广场数据源自建——下掉 veridrop 抓取层，观测改读自家众测快照
- relay/leaderboard.rs → relay/directory.rs：广场成为三份缓存+探针记录的纯读取
  投影（远端配置 / ai-transit / crowd 快照），自有抓取物与 4 份 6h 缓存、
  4 个刷新协调器、LeaderboardKind 协议分榜全删
- 行级观测（TTFT/错误率/名次）公开：decorate_crowd 按 apex join 快照 w24，
  不查共建开关；深数据门禁仍在 crowd_get_snapshot（详情弹窗）
- 名次=实测健康序：TTFT p50 升序→错误率次键→无数据垫底 host 字典序，全行 1..N
- 命令 relay_list_directory 改同步纯读（SWR 后台追新快照）；relay_refresh_directory
  同步刷快照+异步刷 transit；事件载荷去 kind
- maintenance veridrop-directory→directory：+crowd 快照周期刷新，-榜单刷新
- RelayDirectorySite 删 veridrop_host（join 键=apex，无需别名）；卸 scraper 依赖
- 端到端闸换源不降级：零实测数据时每个受管站仍有行；新增 DTO 键形闸
  （camelCase + 不残留换源前死键）
