+++ src-tauri/Cargo.toml
@@ -111,7 +111,6 @@ rustls = "0.23"
 webpki-roots = "1.0"
 rustls-native-certs = "0.8"
 regex = "1.10"
-scraper = "0.27"
 rquickjs = { version = "0.12", features = ["array-buffer", "classes"] }
 thiserror = "2.0"
 anyhow = "1.0"
