--- src-tauri/src/relay/fixtures/veridrop-claude.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html><body>
-  <article class="lb-row official-baseline-row" id="api.anthropic.com">
-    <span class="lb-score-num">100</span>
-  </article>
-  <article class="lb-row" id="bestapi.store" data-impression-domain="bestapi.store" data-impression-surface="leaderboard_top" data-impression-position="14">
-    <div class="lb-main">
-      <h2 class="lb-domain"><a href="/leaderboard/bestapi.store?model=claude">bestapi.store</a></h2>
-      <p class="lb-meta"><strong>20</strong> 次 Claude 检测 · 最近 2026-08-12</p>
-      <div class="lb-protocols">
-        <a class="lb-proto-chip" href="/r/claude-best"><span class="lb-proto-name">Claude</span><span class="lb-proto-score">99</span><span class="lb-proto-verdict">通过</span><span class="lb-proto-count">×20</span><span class="lb-proto-sig">🔐 95%</span></a>
-        <a class="lb-proto-chip" href="/r/openai-best"><span class="lb-proto-name">OpenAI</span><span class="lb-proto-score">95</span><span class="lb-proto-count">×27</span></a>
-      </div>
-      <div><span title="thinking signature 95%">💻 Claude Code / Cursor 编程</span><span title="identity">🧠 复杂推理 / Opus thinking</span></div>
-      <p class="lb-issues">常见问题: <code>pdf</code><code>token_usage</code></p>
-    </div>
-    <div class="lb-score"><span class="lb-score-num">99</span><span class="lb-score-label">Claude 中位</span></div>
-  </article>
-  <article class="lb-row" id="api.790053500.com" data-impression-domain="api.790053500.com" data-impression-surface="leaderboard_all" data-impression-position="72">
-    <div class="lb-main"><h2 class="lb-domain"><a href="/leaderboard/api.790053500.com?model=claude">api.790053500.com</a></h2><p class="lb-meta"><strong>9</strong> 次 Claude 检测 · 最近 2026-08-11</p><div class="lb-protocols"><a class="lb-proto-chip"><span class="lb-proto-name">Claude</span><span class="lb-proto-score">96</span><span class="lb-proto-count">×9</span></a></div></div>
-    <div class="lb-score"><span class="lb-score-num">96</span><span class="lb-score-label">Claude 中位</span></div>
-  </article>
-  <article class="lb-row" id="hidden.example" data-impression-domain="hidden.example" data-impression-surface="leaderboard_all" data-impression-position="73"><span class="lb-score-num">95</span></article>
-</body></html>
