--- src-tauri/src/relay/fixtures/veridrop-detail-wawapii.html
@@ -1,37 +1,0 @@
-<!doctype html>
-<html><body>
-  <section class="hub-hero">
-    <p class="answer-capsule">
-      <strong>wawapii.com 中转站</strong> 在 Veridrop 上累计 <strong>276</strong> 次独立检测，
-      中位分 <strong>100/100</strong>。最近一次检测:2026-08-13。
-    </p>
-  </section>
-  <section class="lb-detail-summary">
-    <div class="lb-detail-score"><span class="lb-detail-score-num">100</span></div>
-    <div class="lb-detail-protocols">
-      <div class="lb-protocols">
-        <a class="lb-proto-chip" href="/r/claude-wawapii">
-          <span class="lb-proto-name">Claude</span>
-          <span class="lb-proto-score">100</span>
-          <span class="lb-proto-verdict">通过</span>
-          <span class="lb-proto-count">×62</span>
-        </a>
-        <a class="lb-proto-chip" href="/r/openai-wawapii">
-          <span class="lb-proto-name">OpenAI</span>
-          <span class="lb-proto-score">100</span>
-          <span class="lb-proto-verdict">通过</span>
-          <span class="lb-proto-count">×213</span>
-        </a>
-        <a class="lb-proto-chip" href="/r/gemini-wawapii">
-          <span class="lb-proto-name">Gemini</span>
-          <span class="lb-proto-score">100</span>
-          <span class="lb-proto-verdict">通过</span>
-          <span class="lb-proto-count">×1</span>
-        </a>
-      </div>
-    </div>
-  </section>
-  <section class="lb-detail-issues">
-    <code>token_usage</code>
-  </section>
-</body></html>
