--- src-tauri/src/relay/fixtures/veridrop-gemini.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html><body>
-  <article class="lb-row" data-impression-domain="gemini.example" data-impression-surface="leaderboard_top" data-impression-position="3">
-    <div class="lb-main">
-      <h2 class="lb-domain"><a href="/leaderboard/gemini.example?model=gemini">gemini.example</a></h2>
-      <p class="lb-meta">16 次 Gemini 检测 · 最近 2026-08-12</p>
-      <div class="lb-protocols">
-        <a class="lb-proto-chip" href="/r/claude-example"><span class="lb-proto-name">Claude</span><span class="lb-proto-score">91</span><span class="lb-proto-count">12</span><span class="lb-proto-verdict">通过</span></a>
-      </div>
-    </div>
-    <span class="lb-score-num">98</span>
-  </article>
-</body></html>
