--- src-tauri/src/relay/fixtures/veridrop-openai.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html><body>
-  <article class="lb-row" data-impression-domain="bestapi.store" data-impression-surface="leaderboard_top" data-impression-position="6">
-    <div class="lb-main">
-      <h2 class="lb-domain"><a href="/leaderboard/bestapi.store?model=openai">bestapi.store</a></h2>
-      <p class="lb-meta">27 次 OpenAI 检测 · 最近 2026-08-12</p>
-      <div class="lb-protocols">
-        <a class="lb-proto-chip" href="/r/openai-best"><span class="lb-proto-name">OpenAI</span><span class="lb-proto-score">95</span><span class="lb-proto-count">27</span><span class="lb-proto-verdict">通过</span></a>
-      </div>
-    </div>
-    <span class="lb-score-num">95</span>
-  </article>
-</body></html>
