--- src-tauri/src/relay/fixtures/veridrop-overall.html
@@ -1,14 +1,0 @@
-<!doctype html>
-<html><body>
-  <article class="lb-row" data-impression-domain="bestapi.store" data-impression-surface="leaderboard_top" data-impression-position="8">
-    <div class="lb-main">
-      <h2 class="lb-domain"><a href="/leaderboard/bestapi.store">bestapi.store</a></h2>
-      <p class="lb-meta">47 次综合检测 · 最近 2026-08-12</p>
-      <div class="lb-protocols">
-        <a class="lb-proto-chip" href="/r/overall-claude"><span class="lb-proto-name">Claude</span><span class="lb-proto-score">99</span><span class="lb-proto-count">20</span></a>
-        <a class="lb-proto-chip" href="/r/overall-openai"><span class="lb-proto-name">OpenAI</span><span class="lb-proto-score">95</span><span class="lb-proto-count">27</span></a>
-      </div>
-    </div>
-    <span class="lb-score-num">95</span>
-  </article>
-</body></html>
