--- src-tauri/src/relay/leaderboard.rs
@@ -1,2359 +1,0 @@
-use scraper::{ElementRef, Html, Selector};
-use serde::{Deserialize, Serialize};
-use std::collections::{BTreeMap, BTreeSet};
-use std::sync::atomic::{AtomicU64, Ordering};
-use std::time::{Duration, Instant};
-
-use crate::error::AppError;
-use crate::maintenance::config::VERIDROP_CACHE_TTL;
-use crate::relay::remote_config::{RelayDirectoryPolicy, RemoteConfig};
-
-const VERIDROP_ORIGIN: &str = "https://veridrop.org";
-const MAX_PAGE_BYTES: usize = 2 * 1024 * 1024;
-const FETCH_TIMEOUT_SECS: u64 = 12;
-const MANAGED_DETAIL_CONCURRENCY: usize = 4;
-const CACHE_SCHEMA_VERSION: u8 = 2;
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "snake_case")]
-pub enum LeaderboardKind {
-    Overall,
-    Claude,
-    #[serde(rename = "openai")]
-    OpenAi,
-    Gemini,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ProtocolScore {
-    pub protocol: String,
-    pub score: u8,
-    pub samples: u32,
-    pub verdict: Option<String>,
-    pub report_url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ParsedLeaderboardItem {
-    pub veridrop_host: String,
-    pub rank: Option<u32>,
-    pub score: u8,
-    pub samples: u32,
-    pub latest_date: String,
-    pub detail_url: String,
-    pub protocol_scores: Vec<ProtocolScore>,
-    pub claude_signature_rate: Option<u8>,
-    pub scenarios: Vec<String>,
-    pub issues: Vec<String>,
-}
-
-/// 广场的一行。veridrop 观测数据 + 站方一手 transit 摘要（**两者都可缺** ——
-/// 未部署公开协议的站没有 transit；不在 veridrop 榜上的受管站连观测四件套
-/// 一起缺，行仍在、徽章不渲染，见 [`apply_policy`] 的合成兜底）。
-///
-/// `Eq` 主动放弃：transit 摘要带 f64，而本结构只参与 `assert_eq!`
-/// （`PartialEq`）——没有 HashMap/HashSet 语义需要 `Eq`。
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelayDirectoryItem {
-    pub site_host: String,
-    /// 站点**身份**：注册域（apex，[`crate::relay::identity::site_domain`] 的产出）。
-    /// 与 `site_host`（真实 host）分工：身份给跨数据源的 join（如实测快照的
-    /// 站点键），真实 host 给链接与取数 —— 别拿一个当另一个用。
-    pub site_domain: String,
-    pub veridrop_host: String,
-    pub display_name: String,
-    pub rank: Option<u32>,
-    /// veridrop 观测四件套可整体缺席：受管站不在 veridrop 榜/详情页时，行由
-    /// [`apply_policy`] 合成，这四个字段全 `None`，前端不渲染对应徽章。
-    /// **不能拿 0 / 空串当「无数据」** —— 0 分、0 样本是真实可能的观测值。
-    pub score: Option<u8>,
-    pub samples: Option<u32>,
-    pub latest_date: Option<String>,
-    pub detail_url: Option<String>,
-    pub protocol_scores: Vec<ProtocolScore>,
-    pub claude_signature_rate: Option<u8>,
-    pub scenarios: Vec<String>,
-    pub issues: Vec<String>,
-    pub entry_url: String,
-    /// 一键登录可用。广场只展示受管站点（见 [`apply_policy`] 的白名单过滤），
-    /// 而受管 = 探针验证过 sub2api / newapi 协议 ⇒ 每一行都该走应用内登录，
-    /// 不再区分「在浏览器打开让用户自己注册」的旧路径。
-    pub auto_add: bool,
-    /// 站方一手 transit 摘要（ai-transit.v1）。`None`/缺席 = 该站没有公开
-    /// 协议数据，前端不渲染徽章。由 [`decorate_transit`] 在**读取路径**合并，
-    /// 不进榜单缓存——两份数据各有各的刷新周期，缓存里存的是 veridrop
-    /// 观测的原始形状。
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub transit: Option<crate::relay::transit::TransitSummary>,
-}
-
-/// 返回给 UI 的榜单。`Eq` 随 [`RelayDirectoryItem`] 一起放弃（f64 摘要），
-/// 只有 `assert_eq!` 消费它。
-#[derive(Debug, Clone, PartialEq, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-pub struct RelayLeaderboard {
-    pub kind: LeaderboardKind,
-    pub items: Vec<RelayDirectoryItem>,
-    pub synced_at: i64,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Deserialize, Serialize)]
-#[serde(rename_all = "camelCase")]
-struct CachedLeaderboard {
-    schema_version: u8,
-    kind: LeaderboardKind,
-    items: Vec<ParsedLeaderboardItem>,
-    managed_hosts: Vec<String>,
-    synced_at: i64,
-}
-
-#[derive(Debug, Clone)]
-pub struct RefreshOutcome {
-    pub leaderboard: RelayLeaderboard,
-    pub updated: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RefreshMode {
-    Force,
-    IfStale,
-}
-
-struct RefreshCoordinator {
-    generation: AtomicU64,
-    slot: tokio::sync::Mutex<RefreshSlot>,
-}
-
-struct RefreshSlot {
-    last_result: Option<Result<RelayLeaderboard, String>>,
-    last_failure_at: Option<Instant>,
-}
-
-impl RefreshCoordinator {
-    const fn new() -> Self {
-        Self {
-            generation: AtomicU64::new(0),
-            slot: tokio::sync::Mutex::const_new(RefreshSlot {
-                last_result: None,
-                last_failure_at: None,
-            }),
-        }
-    }
-
-    async fn run<Fresh, Task, Fut>(
-        &self,
-        mode: RefreshMode,
-        cache_available: bool,
-        fresh: Fresh,
-        task: Task,
-    ) -> Result<RefreshOutcome, AppError>
-    where
-        Fresh: Fn() -> Option<RelayLeaderboard>,
-        Task: FnOnce() -> Fut,
-        Fut: std::future::Future<Output = Result<RelayLeaderboard, AppError>>,
-    {
-        let observed_generation = self.generation.load(Ordering::Acquire);
-        let mut slot = self.slot.lock().await;
-        let generation_changed = self.generation.load(Ordering::Acquire) != observed_generation;
-        let fresh_leaderboard = (mode == RefreshMode::IfStale || generation_changed)
-            .then(fresh)
-            .flatten();
-
-        if mode == RefreshMode::IfStale {
-            if let Some(leaderboard) = fresh_leaderboard.clone() {
-                return Ok(RefreshOutcome {
-                    leaderboard,
-                    updated: false,
-                });
-            }
-        }
-
-        if generation_changed {
-            match slot.last_result.as_ref() {
-                Some(Err(_)) | None => return shared_refresh_result(&slot),
-                Some(Ok(_)) => {
-                    if let Some(leaderboard) = fresh_leaderboard {
-                        return Ok(RefreshOutcome {
-                            leaderboard,
-                            updated: false,
-                        });
-                    }
-                }
-            }
-        }
-
-        if mode == RefreshMode::IfStale
-            && cache_available
-            && slot.last_failure_at.is_some_and(|failed_at| {
-                failed_at.elapsed() < crate::maintenance::config::VERIDROP_RETRY_DELAY
-            })
-        {
-            return shared_refresh_result(&slot);
-        }
-
-        let result = task().await.map_err(|error| error.to_string());
-        slot.last_failure_at = result.as_ref().err().map(|_| Instant::now());
-        slot.last_result = Some(result.clone());
-        self.generation.fetch_add(1, Ordering::Release);
-
-        result
-            .map(|leaderboard| RefreshOutcome {
-                leaderboard,
-                updated: true,
-            })
-            .map_err(AppError::Message)
-    }
-}
-
-impl Default for RefreshCoordinator {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-fn shared_refresh_result(slot: &RefreshSlot) -> Result<RefreshOutcome, AppError> {
-    slot.last_result
-        .clone()
-        .ok_or_else(|| AppError::Message("VeriDrop 刷新状态缺少结果".into()))?
-        .map(|leaderboard| RefreshOutcome {
-            leaderboard,
-            updated: false,
-        })
-        .map_err(AppError::Message)
-}
-
-impl LeaderboardKind {
-    pub const ALL: [Self; 4] = [Self::Overall, Self::Claude, Self::OpenAi, Self::Gemini];
-
-    fn path(self) -> &'static str {
-        match self {
-            Self::Overall => "/leaderboard",
-            Self::Claude => "/leaderboard/claude",
-            Self::OpenAi => "/leaderboard/openai",
-            Self::Gemini => "/leaderboard/gemini",
-        }
-    }
-
-    fn cache_name(self) -> &'static str {
-        match self {
-            Self::Overall => "overall",
-            Self::Claude => "claude",
-            Self::OpenAi => "openai",
-            Self::Gemini => "gemini",
-        }
-    }
-
-    fn protocol_name(self) -> Option<&'static str> {
-        match self {
-            Self::Overall => None,
-            Self::Claude => Some("Claude"),
-            Self::OpenAi => Some("OpenAI"),
-            Self::Gemini => Some("Gemini"),
-        }
-    }
-}
-
-fn selector(value: &str) -> Result<Selector, AppError> {
-    Selector::parse(value)
-        .map_err(|error| AppError::Config(format!("VeriDrop 选择器无效 {value}: {error}")))
-}
-
-fn text(element: ElementRef<'_>) -> String {
-    element.text().collect::<String>().trim().to_string()
-}
-
-fn parse_u32(value: &str) -> Option<u32> {
-    value
-        .chars()
-        .filter(char::is_ascii_digit)
-        .collect::<String>()
-        .parse()
-        .ok()
-}
-
-fn parse_score(value: &str) -> Option<u8> {
-    parse_u32(value).and_then(|value| u8::try_from(value).ok())
-}
-
-fn absolute_url(value: &str) -> Option<String> {
-    let url = url::Url::parse(value)
-        .or_else(|_| url::Url::parse(VERIDROP_ORIGIN).and_then(|base| base.join(value)))
-        .ok()?;
-    (url.scheme() == "https" && url.host_str() == Some("veridrop.org")).then(|| url.to_string())
-}
-
-pub fn parse_page(
-    _kind: LeaderboardKind,
-    html: &str,
-    managed_hosts: &[String],
-) -> Result<Vec<ParsedLeaderboardItem>, AppError> {
-    let document = Html::parse_document(html);
-    let row_selector = selector("article.lb-row[data-impression-domain]")?;
-    let score_selector = selector(".lb-score-num")?;
-    let meta_selector = selector(".lb-meta")?;
-    let detail_selector = selector(".lb-domain a, .lb-detail-link")?;
-    let protocol_selector = selector(".lb-proto-chip")?;
-    let protocol_name_selector = selector(".lb-proto-name")?;
-    let protocol_score_selector = selector(".lb-proto-score")?;
-    let protocol_count_selector = selector(".lb-proto-count")?;
-    let protocol_verdict_selector = selector(".lb-proto-verdict")?;
-    let signature_selector = selector(".lb-proto-sig")?;
-    let issue_selector = selector(".lb-issues code")?;
-    let scenario_selector = selector(".lb-main > div > span[title]")?;
-    let latest_date_pattern =
-        regex::Regex::new(r"最近\s+(\d{4}-\d{2}-\d{2})").expect("static regex");
-
-    let mut items = Vec::new();
-    for row in document.select(&row_selector) {
-        let host = row
-            .value()
-            .attr("data-impression-domain")
-            .map(crate::relay::identity::request_host)
-            .filter(|host| !host.is_empty())
-            .ok_or_else(|| AppError::Config("VeriDrop 榜单行缺少域名".into()))?;
-        let top = row.value().attr("data-impression-surface") == Some("leaderboard_top");
-        if !top && !managed_hosts.iter().any(|managed| managed == &host) {
-            continue;
-        }
-        let rank = row
-            .value()
-            .attr("data-impression-position")
-            .and_then(parse_u32)
-            .ok_or_else(|| AppError::Config(format!("VeriDrop 榜单行 {host} 缺少排名")))?;
-        let score = row
-            .select(&score_selector)
-            .next()
-            .and_then(|element| parse_score(&text(element)))
-            .filter(|score| *score <= 100)
-            .ok_or_else(|| AppError::Config(format!("VeriDrop 榜单行 {host} 分数无效")))?;
-        let meta = row
-            .select(&meta_selector)
-            .next()
-            .map(text)
-            .unwrap_or_default();
-        let samples = parse_u32(meta.split('次').next().unwrap_or_default()).unwrap_or(0);
-        let latest_date = latest_date_pattern
-            .captures(&meta)
-            .and_then(|captures| captures.get(1))
-            .map(|value| value.as_str().to_string())
-            .unwrap_or_default();
-        let detail_url = row
-            .select(&detail_selector)
-            .filter_map(|element| element.value().attr("href"))
-            .find_map(absolute_url)
-            .ok_or_else(|| AppError::Config(format!("VeriDrop 榜单行 {host} 缺少详情链接")))?;
-
-        let mut protocol_scores = Vec::new();
-        let mut claude_signature_rate = None;
-        for protocol in row.select(&protocol_selector) {
-            let Some(name) = protocol.select(&protocol_name_selector).next().map(text) else {
-                continue;
-            };
-            let Some(protocol_score) = protocol
-                .select(&protocol_score_selector)
-                .next()
-                .and_then(|element| parse_score(&text(element)))
-                .filter(|score| *score <= 100)
-            else {
-                continue;
-            };
-            let protocol_samples = protocol
-                .select(&protocol_count_selector)
-                .next()
-                .and_then(|element| parse_u32(&text(element)))
-                .unwrap_or(0);
-            let verdict = protocol.select(&protocol_verdict_selector).next().map(text);
-            let report_url = protocol.value().attr("href").and_then(absolute_url);
-            if name == "Claude" {
-                claude_signature_rate = protocol
-                    .select(&signature_selector)
-                    .next()
-                    .and_then(|element| parse_score(&text(element)));
-            }
-            protocol_scores.push(ProtocolScore {
-                protocol: name,
-                score: protocol_score,
-                samples: protocol_samples,
-                verdict,
-                report_url,
-            });
-        }
-        let scenarios = row
-            .select(&scenario_selector)
-            .map(text)
-            .map(|value| {
-                value
-                    .trim_start_matches(|character: char| !character.is_alphanumeric())
-                    .trim()
-                    .to_string()
-            })
-            .filter(|value| !value.is_empty())
-            .collect();
-        let issues = row.select(&issue_selector).map(text).collect();
-        items.push(ParsedLeaderboardItem {
-            veridrop_host: host,
-            rank: Some(rank),
-            score,
-            samples,
-            latest_date,
-            detail_url,
-            protocol_scores,
-            claude_signature_rate,
-            scenarios,
-            issues,
-        });
-    }
-    Ok(items)
-}
-
-fn parse_detail_page(
-    kind: LeaderboardKind,
-    host: &str,
-    html: &str,
-) -> Result<Option<ParsedLeaderboardItem>, AppError> {
-    let document = Html::parse_document(html);
-    let score_selector = selector(".lb-detail-score-num")?;
-    let protocol_selector = selector(".lb-detail-protocols .lb-proto-chip")?;
-    let protocol_name_selector = selector(".lb-proto-name")?;
-    let protocol_score_selector = selector(".lb-proto-score")?;
-    let protocol_count_selector = selector(".lb-proto-count")?;
-    let protocol_verdict_selector = selector(".lb-proto-verdict")?;
-    let signature_selector = selector(".lb-proto-sig")?;
-    let issue_selector = selector(".lb-detail-issues code")?;
-    let summary_selector = selector(".answer-capsule")?;
-    let latest_date_pattern =
-        regex::Regex::new(r"最近一次检测[:：]\s*(\d{4}-\d{2}-\d{2})").expect("static regex");
-    let overall_samples_pattern =
-        regex::Regex::new(r"累计\s*(\d+)\s*次独立检测").expect("static regex");
-
-    let mut protocol_scores = Vec::new();
-    let mut claude_signature_rate = None;
-    for protocol in document.select(&protocol_selector) {
-        let Some(name) = protocol.select(&protocol_name_selector).next().map(text) else {
-            continue;
-        };
-        let Some(score) = protocol
-            .select(&protocol_score_selector)
-            .next()
-            .and_then(|element| parse_score(&text(element)))
-            .filter(|score| *score <= 100)
-        else {
-            continue;
-        };
-        let samples = protocol
-            .select(&protocol_count_selector)
-            .next()
-            .and_then(|element| parse_u32(&text(element)))
-            .unwrap_or(0);
-        let verdict = protocol.select(&protocol_verdict_selector).next().map(text);
-        let report_url = protocol.value().attr("href").and_then(absolute_url);
-        if name == "Claude" {
-            claude_signature_rate = protocol
-                .select(&signature_selector)
-                .next()
-                .and_then(|element| parse_score(&text(element)));
-        }
-        protocol_scores.push(ProtocolScore {
-            protocol: name,
-            score,
-            samples,
-            verdict,
-            report_url,
-        });
-    }
-
-    let (score, samples) = match kind.protocol_name() {
-        Some(protocol_name) => {
-            let Some(protocol) = protocol_scores
-                .iter()
-                .find(|protocol| protocol.protocol == protocol_name)
-            else {
-                return Ok(None);
-            };
-            (protocol.score, protocol.samples)
-        }
-        None => {
-            let Some(score) = document
-                .select(&score_selector)
-                .next()
-                .and_then(|element| parse_score(&text(element)))
-                .filter(|score| *score <= 100)
-            else {
-                return Ok(None);
-            };
-            let summary = document
-                .select(&summary_selector)
-                .next()
-                .map(text)
-                .unwrap_or_default();
-            let samples = overall_samples_pattern
-                .captures(&summary)
-                .and_then(|captures| captures.get(1))
-                .and_then(|value| value.as_str().parse().ok())
-                .unwrap_or(0);
-            (score, samples)
-        }
-    };
-    let summary = document
-        .select(&summary_selector)
-        .next()
-        .map(text)
-        .unwrap_or_default();
-    let latest_date = latest_date_pattern
-        .captures(&summary)
-        .and_then(|captures| captures.get(1))
-        .map(|value| value.as_str().to_string())
-        .unwrap_or_default();
-    let host = crate::relay::identity::request_host(host);
-    if host.is_empty() {
-        return Ok(None);
-    }
-
-    Ok(Some(ParsedLeaderboardItem {
-        detail_url: format!("{VERIDROP_ORIGIN}/leaderboard/{host}"),
-        veridrop_host: host,
-        rank: None,
-        score,
-        samples,
-        latest_date,
-        protocol_scores,
-        claude_signature_rate,
-        scenarios: vec![],
-        issues: document.select(&issue_selector).map(text).collect(),
-    }))
-}
-
-fn normalized_policy(policy: &RelayDirectoryPolicy) -> RelayDirectoryPolicy {
-    RelayDirectoryPolicy {
-        blocked_hosts: policy
-            .blocked_hosts
-            .iter()
-            .map(|host| crate::relay::identity::request_host(host))
-            .filter(|host| !host.is_empty())
-            .collect(),
-        sites: policy
-            .sites
-            .iter()
-            .map(|(host, site)| (crate::relay::identity::request_host(host), site.clone()))
-            .filter(|(host, _)| !host.is_empty())
-            .collect(),
-    }
-}
-
-fn managed_veridrop_hosts(config: &RemoteConfig) -> Vec<String> {
-    let policy = normalized_policy(&config.relay_directory);
-    let mut blocked: BTreeSet<_> = policy.blocked_hosts.iter().cloned().collect();
-    let mut aliases = BTreeMap::new();
-    for (loongport_host, site) in &policy.sites {
-        let veridrop_host = site
-            .veridrop_host
-            .as_deref()
-            .map(crate::relay::identity::request_host)
-            .filter(|host| !host.is_empty())
-            .unwrap_or_else(|| loongport_host.clone());
-        aliases.insert(loongport_host.clone(), veridrop_host.clone());
-        aliases.insert(veridrop_host.clone(), veridrop_host.clone());
-        if blocked.contains(loongport_host) || blocked.contains(&veridrop_host) {
-            blocked.insert(loongport_host.clone());
-            blocked.insert(veridrop_host);
-        }
-    }
-
-    let mut candidates = BTreeSet::new();
-    for sponsor in &config.sponsors {
-        candidates.insert(crate::relay::identity::request_host(&sponsor.site_origin));
-    }
-    candidates.extend(
-        config
-            .aff_codes
-            .keys()
-            .map(|host| crate::relay::identity::request_host(host)),
-    );
-    candidates.extend(
-        config
-            .promo_codes
-            .keys()
-            .map(|host| crate::relay::identity::request_host(host)),
-    );
-    for (loongport_host, site) in &policy.sites {
-        candidates.insert(
-            site.veridrop_host
-                .as_deref()
-                .map(crate::relay::identity::request_host)
-                .filter(|host| !host.is_empty())
-                .unwrap_or_else(|| loongport_host.clone()),
-        );
-    }
-    candidates
-        .into_iter()
-        .filter(|host| !host.is_empty() && !blocked.contains(host))
-        .map(|host| aliases.get(&host).cloned().unwrap_or(host))
-        .filter(|host| !blocked.contains(host))
-        .collect::<BTreeSet<_>>()
-        .into_iter()
-        .collect()
-}
-
-/// 受管站点的**真实 origin** host 全集（sponsors ∪ aff_codes ∪ promo_codes ∪
-/// relay_directory，统一归一后去 blocked）。
-///
-/// ⭐ 三份清单的**唯一源**（wawapi.top 实测踩出的洞：aff 名单里的站进了广场，
-/// 导入闸却只认 relay_directory ⇒ 广场显示、点接入被拒）。广场曝光
-/// （`managed_veridrop_hosts` 是这份在 veridrop 空间的别名视图）、目录导入闸的
-/// 回落匹配（`commands::relay::directory_entry_source`）、探针名单
-/// （`refresh_site_probes_for_directory`）全部从这份派生 —— 任何一份单独收窄，
-/// 就会重现「广场显示但点不进 / 探针名单漏探」，三处必须同宽。
-pub(crate) fn managed_site_hosts(config: &RemoteConfig) -> Vec<String> {
-    let policy = normalized_policy(&config.relay_directory);
-    let blocked: BTreeSet<_> = policy.blocked_hosts.iter().cloned().collect();
-    let mut hosts: BTreeSet<String> = policy.sites.into_keys().collect();
-    for sponsor in &config.sponsors {
-        hosts.insert(crate::relay::identity::request_host(&sponsor.site_origin));
-    }
-    hosts.extend(
-        config
-            .aff_codes
-            .keys()
-            .map(|host| crate::relay::identity::request_host(host)),
-    );
-    hosts.extend(
-        config
-            .promo_codes
-            .keys()
-            .map(|host| crate::relay::identity::request_host(host)),
-    );
-    hosts
-        .into_iter()
-        .filter(|host| !host.is_empty() && !blocked.contains(host))
-        .collect()
-}
-
-/// 四个榜单缓存里出现过的 veridrop host 全集 —— 漏斗第 1/2 层的对照面。
-fn cached_veridrop_hosts() -> BTreeSet<String> {
-    LeaderboardKind::ALL
-        .into_iter()
-        .filter_map(read_cache)
-        .flat_map(|cached| cached.items)
-        .map(|item| item.veridrop_host)
-        .collect()
-}
-
-/// 目录漏斗的周期收尾：**三层过滤逐层留痕** + 受管站探针落盘。
-///
-/// 由 `maintenance` 的 veridrop-directory 定时任务在配置与榜单刷新之后调用，
-/// 每个周期（6 小时）输出三条 INFO：
-///
-/// 1. `whitelist_filter` —— 榜单在、白名单不在（未收录，不展示）；
-/// 2. `leaderboard_missing` —— 白名单在、榜单没有行（无排名数据但仍可展示）；
-/// 3. `probe`（逐站一条）—— 探针三分类 + 连续 miss 计数 + 是否仍曝光。
-///
-/// 下一轮排查「这个站为什么不在广场」时，逐层翻日志就能回答。
-pub(crate) async fn refresh_site_probes_for_directory() {
-    let config = crate::relay::remote_config::load_cached().unwrap_or_default();
-    let managed_hosts = managed_site_hosts(&config);
-    if managed_hosts.is_empty() {
-        log::info!(
-            "{}",
-            crate::diagnostics::DiagnosticEvent::new("relay.directory_funnel", "skipped")
-                .field("reason", "no_managed_sites")
-        );
-        return;
-    }
-
-    let managed_veridrop: BTreeSet<_> = managed_veridrop_hosts(&config).into_iter().collect();
-    let cached_hosts = cached_veridrop_hosts();
-    let not_managed: Vec<_> = cached_hosts
-        .difference(&managed_veridrop)
-        .cloned()
-        .collect();
-    let missing_from_board: Vec<_> = managed_veridrop
-        .difference(&cached_hosts)
-        .cloned()
-        .collect();
-    log::info!(
-        "{}",
-        crate::diagnostics::DiagnosticEvent::new("relay.directory_funnel", "whitelist_filter")
-            .field("dropped", not_managed.join(","))
-            .field("dropped_count", not_managed.len())
-    );
-    log::info!(
-        "{}",
-        crate::diagnostics::DiagnosticEvent::new("relay.directory_funnel", "leaderboard_missing")
-            .field("hosts", missing_from_board.join(","))
-            .field("count", missing_from_board.len())
-    );
-
-    let origins = managed_hosts
-        .iter()
-        .map(|host| format!("https://{host}"))
-        .collect::<Vec<_>>();
-    for record in crate::relay::site_probe::probe_and_record(&origins).await {
-        log::info!(
-            "{}",
-            crate::diagnostics::DiagnosticEvent::new("relay.directory_funnel", "probe")
-                .field("host", record.host)
-                .field("verdict", format!("{:?}", record.verdict))
-                .field("backend", format!("{:?}", record.backend))
-                .field("consecutive_panel_misses", record.consecutive_panel_misses)
-                .field("exposed", record.exposed)
-                .field("detail", record.detail)
-        );
-    }
-}
-
-pub fn apply_policy(
-    parsed: Vec<ParsedLeaderboardItem>,
-    config: &RemoteConfig,
-    kind: LeaderboardKind,
-) -> Vec<RelayDirectoryItem> {
-    let policy = normalized_policy(&config.relay_directory);
-    let blocked: BTreeSet<_> = policy.blocked_hosts.iter().cloned().collect();
-    let mut aliases = BTreeMap::new();
-    for (loongport_host, site) in &policy.sites {
-        let value = (loongport_host.clone(), site.clone());
-        aliases.insert(loongport_host.clone(), value.clone());
-        let veridrop_host = site
-            .veridrop_host
-            .as_deref()
-            .map(crate::relay::identity::request_host)
-            .filter(|host| !host.is_empty())
-            .unwrap_or_else(|| loongport_host.clone());
-        aliases.insert(veridrop_host, value);
-    }
-    let mut seen_sites = BTreeSet::new();
-    // 展示白名单：只有受管站点（sponsors / aff / promo / relay_directory 声明，
-    // 且未被 block）可以进广场。VeriDrop 榜单只作为这些站点的**观测数据来源**
-    // （排名 / 评分 / 样本），不是展示集合的所有者 —— 未受管的站点（包括 Top 60）
-    // 没经过「探针可通 + 一键登录可用」的收录验证，展示出来就是给用户一个
-    // 点了大概率失败的按钮。
-    let managed: BTreeSet<_> = managed_veridrop_hosts(config).into_iter().collect();
-
-    let mut items: Vec<RelayDirectoryItem> = parsed
-        .into_iter()
-        .filter_map(|item| {
-            let (site_host, override_site) = aliases
-                .get(&item.veridrop_host)
-                .cloned()
-                .map(|(host, site)| (host, Some(site)))
-                .unwrap_or_else(|| (item.veridrop_host.clone(), None));
-            if blocked.contains(&site_host) || blocked.contains(&item.veridrop_host) {
-                return None;
-            }
-            if !managed.contains(&item.veridrop_host) {
-                return None;
-            }
-            if !seen_sites.insert(site_host.clone()) {
-                return None;
-            }
-            let auto_add = true;
-            let entry_url = override_site
-                .as_ref()
-                .and_then(|site| site.entry_url.clone())
-                .filter(|url| url::Url::parse(url).is_ok_and(|url| url.scheme() == "https"))
-                .unwrap_or_else(|| format!("https://{site_host}"));
-            let display_name = override_site
-                .and_then(|site| site.display_name)
-                .filter(|name| !name.trim().is_empty())
-                .unwrap_or_else(|| site_host.clone());
-            let site_domain = crate::relay::identity::site_domain(&site_host);
-            Some(RelayDirectoryItem {
-                site_host,
-                site_domain,
-                veridrop_host: item.veridrop_host,
-                display_name,
-                rank: item.rank,
-                score: Some(item.score),
-                samples: Some(item.samples),
-                latest_date: Some(item.latest_date),
-                detail_url: Some(item.detail_url),
-                protocol_scores: item.protocol_scores,
-                claude_signature_rate: item.claude_signature_rate,
-                scenarios: item.scenarios,
-                issues: item.issues,
-                entry_url,
-                auto_add,
-                transit: None,
-            })
-        })
-        .collect();
-
-    // 白名单是展示的**充分条件**（2026-09-06 维护者拍板）：受管站不在 veridrop
-    // 榜/详情页 ⇒ 缺的只是观测数据，不是行本身 —— 手动添加入口已删的当下，
-    // 广场行是用户接入受管站的唯一入口。仅在总榜合成：协议分榜是「该协议
-    // 有观测数据的站」的排序视图，混入无数据行会稀释它的语义。veridrop 逐步
-    // 被本地自测指标替换后，观测缺席只会更常见，这条兜底就是广场行源的底座。
-    if kind == LeaderboardKind::Overall {
-        items.extend(synthesized_rows_for_missing_sites(config, &policy, &items));
-    }
-    items
-}
-
-/// 给「白名单里有、veridrop 里一行都没有」的受管站合成无数据行。
-///
-/// 行的存在性由**受管名单**（[`managed_site_hosts`]，四张表的并集）决定，
-/// veridrop 只提供观测数据 —— `refresh_site_probes_for_directory` 里
-/// `leaderboard_missing` 那层日志早就宣示了「无排名数据但仍可展示」，本函数
-/// 是它的兑现。去重按**注册域**：sponsors 与 relay_directory 常各录一形
-/// （www. vs 裸域），同一 apex 是一站一行，不是两行。
-fn synthesized_rows_for_missing_sites(
-    config: &RemoteConfig,
-    policy: &RelayDirectoryPolicy,
-    existing: &[RelayDirectoryItem],
-) -> Vec<RelayDirectoryItem> {
-    let mut seen_domains: BTreeSet<String> = existing
-        .iter()
-        .map(|item| item.site_domain.clone())
-        .collect();
-    managed_site_hosts(config)
-        .into_iter()
-        .filter_map(|site_host| {
-            let site_domain = crate::relay::identity::site_domain(&site_host);
-            if !seen_domains.insert(site_domain.clone()) {
-                return None;
-            }
-            let site = policy.sites.get(&site_host);
-            let display_name = site
-                .and_then(|site| site.display_name.as_deref())
-                .filter(|name| !name.trim().is_empty())
-                .map(str::to_string)
-                .unwrap_or_else(|| site_host.clone());
-            let veridrop_host = site
-                .and_then(|site| site.veridrop_host.as_deref())
-                .map(crate::relay::identity::request_host)
-                .filter(|host| !host.is_empty())
-                .unwrap_or_else(|| site_host.clone());
-            let entry_url = site
-                .and_then(|site| site.entry_url.as_deref())
-                .filter(|url| url::Url::parse(url).is_ok_and(|url| url.scheme() == "https"))
-                .map(str::to_string)
-                .unwrap_or_else(|| format!("https://{site_host}"));
-            Some(RelayDirectoryItem {
-                site_host,
-                site_domain,
-                veridrop_host,
-                display_name,
-                rank: None,
-                score: None,
-                samples: None,
-                latest_date: None,
-                detail_url: None,
-                protocol_scores: Vec::new(),
-                claude_signature_rate: None,
-                scenarios: Vec::new(),
-                issues: Vec::new(),
-                entry_url,
-                auto_add: true,
-                transit: None,
-            })
-        })
-        .collect()
-}
-
-fn cache_path(kind: LeaderboardKind) -> std::path::PathBuf {
-    crate::config::get_home_dir()
-        .join(crate::config::APP_DIR_NAME)
-        .join(format!("veridrop-{}.json", kind.cache_name()))
-}
-
-fn read_cache(kind: LeaderboardKind) -> Option<CachedLeaderboard> {
-    let path = cache_path(kind);
-    let metadata = std::fs::metadata(&path).ok()?;
-    if metadata.len() > MAX_PAGE_BYTES as u64 {
-        return None;
-    }
-    let cached: CachedLeaderboard = serde_json::from_slice(&std::fs::read(path).ok()?).ok()?;
-    if cached.schema_version != CACHE_SCHEMA_VERSION || cached.kind != kind {
-        return None;
-    }
-    Some(cached)
-}
-
-fn apply_policy_to_cached(cached: CachedLeaderboard, config: &RemoteConfig) -> RelayLeaderboard {
-    RelayLeaderboard {
-        kind: cached.kind,
-        items: decorate_transit(renumber_ranks_by_score(apply_probe_gate(apply_policy(
-            cached.items,
-            config,
-            cached.kind,
-        )))),
-        synced_at: cached.synced_at,
-    }
-}
-
-/// 把站方一手 transit 摘要合并进广场行（join 键 = 归一后的 site host）。
-///
-/// 只在**读取路径**调用（缓存读与刷新返回两处），榜单缓存里不存它——
-/// transit 有自己的 6 小时刷新周期，写进榜单缓存会让两份数据的过期
-/// 语义咬在一起（榜单没刷新时 transit 也跟着冻结）。
-fn decorate_transit(mut items: Vec<RelayDirectoryItem>) -> Vec<RelayDirectoryItem> {
-    let summaries = crate::relay::transit::summaries();
-    if summaries.is_empty() {
-        return items;
-    }
-    for item in &mut items {
-        item.transit = summaries.get(&item.site_host).cloned();
-    }
-    items
-}
-
-/// 曝光闸：白名单（`apply_policy`）之后再过一遍**探针健康**——连续多轮「协议认不出」
-/// 的站从广场摘掉。见 [`crate::relay::site_probe`] 的三分类（网络失败不摘）。
-fn apply_probe_gate(items: Vec<RelayDirectoryItem>) -> Vec<RelayDirectoryItem> {
-    let store = crate::relay::site_probe::SiteProbeStore::load();
-    filter_probe_gated(items, &store)
-}
-
-/// `apply_probe_gate` 的纯函数核——测试直接喂 store，不碰磁盘。
-fn filter_probe_gated(
-    items: Vec<RelayDirectoryItem>,
-    store: &crate::relay::site_probe::SiteProbeStore,
-) -> Vec<RelayDirectoryItem> {
-    items
-        .into_iter()
-        .filter(|item| store.should_expose(&item.site_host))
-        .collect()
-}
-
-/// 广场最终呈现序：按 veridrop **评分倒排**并重排名次（1..N）。
-///
-/// 白名单与探针闸会摘掉中间名次的站（第 2 名直接跳第 7 名），veridrop 原始名次
-/// 对用户就是乱序。重排后 `rank` = **本广场内的位置**，评分 / 样本数仍是
-/// veridrop 原值 —— 用户看到的既是连续名次，也不丢失「凭什么排这」的依据。
-///
-/// 排序键：评分降序（无评分的合成行排在一切有评分者之后），同分时 veridrop
-/// 原名次小者在前（无原名次的站 —— 靠 detail 页回填的那种 —— 排在同分有名次者
-/// 之后），仍并列按 host 字典序钉死顺序（快照测试友好，两次渲染不跳行）。
-fn renumber_ranks_by_score(mut items: Vec<RelayDirectoryItem>) -> Vec<RelayDirectoryItem> {
-    items.sort_by(|a, b| {
-        b.score
-            .cmp(&a.score)
-            .then_with(|| match (a.rank, b.rank) {
-                (Some(left), Some(right)) => left.cmp(&right),
-                (Some(_), None) => std::cmp::Ordering::Less,
-                (None, Some(_)) => std::cmp::Ordering::Greater,
-                (None, None) => std::cmp::Ordering::Equal,
-            })
-            .then_with(|| a.site_host.cmp(&b.site_host))
-    });
-    for (index, item) in items.iter_mut().enumerate() {
-        item.rank = Some(index as u32 + 1);
-    }
-    items
-}
-
-fn write_cache(leaderboard: &CachedLeaderboard) -> Result<(), AppError> {
-    let path = cache_path(leaderboard.kind);
-    let bytes =
-        serde_json::to_vec(leaderboard).map_err(|source| AppError::JsonSerialize { source })?;
-    crate::config::atomic_write(&path, &bytes)
-}
-
-fn is_fresh_at(synced_at: i64, now: i64) -> bool {
-    now.saturating_sub(synced_at) < VERIDROP_CACHE_TTL.as_secs() as i64
-}
-
-fn is_fresh_for(cached: &CachedLeaderboard, now: i64, config: &RemoteConfig) -> bool {
-    is_fresh_at(cached.synced_at, now) && cached.managed_hosts == managed_veridrop_hosts(config)
-}
-
-pub fn read_cached(kind: LeaderboardKind) -> Result<Option<RelayLeaderboard>, AppError> {
-    let config = crate::relay::remote_config::load_cached().unwrap_or_default();
-    Ok(read_cache(kind).map(|cached| apply_policy_to_cached(cached, &config)))
-}
-
-pub fn is_cache_fresh(kind: LeaderboardKind, now: i64) -> bool {
-    let config = crate::relay::remote_config::load_cached().unwrap_or_default();
-    read_cache(kind).is_some_and(|cached| is_fresh_for(&cached, now, &config))
-}
-
-async fn fetch_html(client: &reqwest::Client, url: &str) -> Result<Option<String>, AppError> {
-    use futures::StreamExt;
-    let response = client
-        .get(url)
-        .send()
-        .await
-        .map_err(|error| AppError::Config(format!("拉取 VeriDrop 页面失败: {error}")))?;
-    if response.status() == reqwest::StatusCode::NOT_FOUND {
-        return Ok(None);
-    }
-    let response = response
-        .error_for_status()
-        .map_err(|error| AppError::Config(format!("VeriDrop 页面响应失败: {error}")))?;
-    if response
-        .content_length()
-        .is_some_and(|length| length > MAX_PAGE_BYTES as u64)
-    {
-        return Err(AppError::Config("VeriDrop 榜单体积异常".into()));
-    }
-    let mut bytes = Vec::new();
-    let mut stream = response.bytes_stream();
-    while let Some(chunk) = stream.next().await {
-        bytes.extend_from_slice(
-            &chunk.map_err(|error| AppError::Config(format!("读取 VeriDrop 榜单失败: {error}")))?,
-        );
-        if bytes.len() > MAX_PAGE_BYTES {
-            return Err(AppError::Config("VeriDrop 榜单超过体积上限".into()));
-        }
-    }
-    String::from_utf8(bytes)
-        .map(Some)
-        .map_err(|error| AppError::Config(format!("VeriDrop 页面不是 UTF-8: {error}")))
-}
-
-async fn fetch_live_source_with(
-    client: &reqwest::Client,
-    origin: &str,
-    kind: LeaderboardKind,
-    managed_hosts: &[String],
-) -> Result<Vec<ParsedLeaderboardItem>, AppError> {
-    use futures::StreamExt;
-
-    let leaderboard_url = format!("{origin}{}", kind.path());
-    let html = fetch_html(client, &leaderboard_url)
-        .await?
-        .ok_or_else(|| AppError::Config("VeriDrop 榜单不存在".into()))?;
-    let mut items = parse_page(kind, &html, managed_hosts)?;
-    let present: BTreeSet<_> = items
-        .iter()
-        .map(|item| item.veridrop_host.clone())
-        .collect();
-    let missing: Vec<_> = managed_hosts
-        .iter()
-        .filter(|host| !present.contains(*host))
-        .cloned()
-        .collect();
-    let mut details = futures::stream::iter(missing.into_iter().map(|host| async move {
-        let detail_url = format!("{origin}/leaderboard/{host}");
-        let Some(html) = fetch_html(client, &detail_url).await? else {
-            return Ok(None);
-        };
-        parse_detail_page(kind, &host, &html)
-    }))
-    .buffered(MANAGED_DETAIL_CONCURRENCY);
-    while let Some(detail) = details.next().await {
-        if let Some(item) = detail? {
-            items.push(item);
-        }
-    }
-    Ok(items)
-}
-
-#[cfg(test)]
-async fn fetch_live_source(
-    kind: LeaderboardKind,
-    managed_hosts: &[String],
-) -> Result<Vec<ParsedLeaderboardItem>, AppError> {
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(FETCH_TIMEOUT_SECS))
-        .user_agent("LoongPort/relay-directory")
-        .build()
-        .map_err(|error| AppError::Config(format!("创建 VeriDrop 请求失败: {error}")))?;
-    fetch_live_source_with(&client, VERIDROP_ORIGIN, kind, managed_hosts).await
-}
-
-static OVERALL_REFRESH: RefreshCoordinator = RefreshCoordinator::new();
-static CLAUDE_REFRESH: RefreshCoordinator = RefreshCoordinator::new();
-static OPENAI_REFRESH: RefreshCoordinator = RefreshCoordinator::new();
-static GEMINI_REFRESH: RefreshCoordinator = RefreshCoordinator::new();
-
-fn refresh_coordinator(kind: LeaderboardKind) -> &'static RefreshCoordinator {
-    match kind {
-        LeaderboardKind::Overall => &OVERALL_REFRESH,
-        LeaderboardKind::Claude => &CLAUDE_REFRESH,
-        LeaderboardKind::OpenAi => &OPENAI_REFRESH,
-        LeaderboardKind::Gemini => &GEMINI_REFRESH,
-    }
-}
-
-async fn refresh_with(
-    client: &reqwest::Client,
-    origin: &str,
-    kind: LeaderboardKind,
-    config: &RemoteConfig,
-) -> Result<RelayLeaderboard, AppError> {
-    let parsed =
-        fetch_live_source_with(client, origin, kind, &managed_veridrop_hosts(config)).await?;
-    let items = apply_policy(parsed.clone(), config, kind);
-    if items.is_empty() {
-        return Err(AppError::Config("VeriDrop 榜单没有可展示站点".into()));
-    }
-    let synced_at = chrono::Utc::now().timestamp();
-    write_cache(&CachedLeaderboard {
-        schema_version: CACHE_SCHEMA_VERSION,
-        kind,
-        items: parsed,
-        managed_hosts: managed_veridrop_hosts(config),
-        synced_at,
-    })?;
-    Ok(RelayLeaderboard {
-        kind,
-        // 返回给 UI 的这份过探针闸并重排名次；缓存里存的始终是未过滤的 parsed，
-        // 摘除闸解除后无需重新抓取即可恢复展示（名次在读取时重算，天然跟上）。
-        items: decorate_transit(renumber_ranks_by_score(apply_probe_gate(items))),
-        synced_at,
-    })
-}
-
-async fn perform_refresh(kind: LeaderboardKind) -> Result<RelayLeaderboard, AppError> {
-    let config = crate::relay::remote_config::load_cached().unwrap_or_default();
-    let client = reqwest::Client::builder()
-        .timeout(Duration::from_secs(FETCH_TIMEOUT_SECS))
-        .user_agent("LoongPort/relay-directory")
-        .build()
-        .map_err(|error| AppError::Config(format!("创建 VeriDrop 请求失败: {error}")))?;
-    refresh_with(&client, VERIDROP_ORIGIN, kind, &config).await
-}
-
-fn read_fresh(kind: LeaderboardKind) -> Option<RelayLeaderboard> {
-    let config = crate::relay::remote_config::load_cached().unwrap_or_default();
-    let cached = read_cache(kind)?;
-    is_fresh_for(&cached, chrono::Utc::now().timestamp(), &config)
-        .then(|| apply_policy_to_cached(cached, &config))
-}
-
-pub async fn refresh(kind: LeaderboardKind) -> Result<RefreshOutcome, AppError> {
-    refresh_coordinator(kind)
-        .run(
-            RefreshMode::Force,
-            false,
-            || read_fresh(kind),
-            || perform_refresh(kind),
-        )
-        .await
-}
-
-pub async fn refresh_if_stale(kind: LeaderboardKind) -> Result<RefreshOutcome, AppError> {
-    let cache_available = read_cache(kind).is_some();
-    refresh_coordinator(kind)
-        .run(
-            RefreshMode::IfStale,
-            cache_available,
-            || read_fresh(kind),
-            || perform_refresh(kind),
-        )
-        .await
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::{response::Html, routing::get, Router};
-    use serial_test::serial;
-
-    struct TestHomeGuard(Option<std::ffi::OsString>);
-
-    impl TestHomeGuard {
-        fn set(path: &std::path::Path) -> Self {
-            let previous = std::env::var_os("CC_SWITCH_TEST_HOME");
-            std::env::set_var("CC_SWITCH_TEST_HOME", path);
-            Self(previous)
-        }
-    }
-
-    impl Drop for TestHomeGuard {
-        fn drop(&mut self) {
-            match self.0.take() {
-                Some(previous) => std::env::set_var("CC_SWITCH_TEST_HOME", previous),
-                None => std::env::remove_var("CC_SWITCH_TEST_HOME"),
-            }
-        }
-    }
-
-    fn config_with_directory() -> RemoteConfig {
-        RemoteConfig {
-            relay_directory: RelayDirectoryPolicy {
-                blocked_hosts: vec!["blocked.example".into()],
-                sites: BTreeMap::from([
-                    (
-                        "790053500.com".into(),
-                        crate::relay::remote_config::RelayDirectorySite {
-                            veridrop_host: Some("api.790053500.com".into()),
-                            entry_url: Some("https://790053500.com/keys".into()),
-                            purchase_url: None,
-                            usage_url: None,
-                            display_name: Some("鑫旺".into()),
-                        },
-                    ),
-                    (
-                        "blocked.example".into(),
-                        crate::relay::remote_config::RelayDirectorySite {
-                            veridrop_host: None,
-                            entry_url: None,
-                            purchase_url: None,
-                            usage_url: None,
-                            display_name: None,
-                        },
-                    ),
-                ]),
-            },
-            ..RemoteConfig::default()
-        }
-    }
-
-    #[test]
-    fn parse_certified_rows_and_required_metadata() {
-        let page = parse_page(
-            LeaderboardKind::Claude,
-            include_str!("fixtures/veridrop-claude.html"),
-            &["api.790053500.com".into()],
-        )
-        .expect("parse fixture");
-
-        assert_eq!(page.len(), 2);
-        assert_eq!(page[0].veridrop_host, "bestapi.store");
-        assert_eq!(page[0].rank, Some(14));
-        assert_eq!(page[0].score, 99);
-        assert_eq!(page[0].samples, 20);
-        assert_eq!(page[0].latest_date, "2026-08-12");
-        assert_eq!(page[0].claude_signature_rate, Some(95));
-        assert!(page[0]
-            .scenarios
-            .contains(&"Claude Code / Cursor 编程".to_string()));
-        assert_eq!(page[1].veridrop_host, "api.790053500.com");
-        assert_eq!(
-            page[1].rank,
-            Some(72),
-            "managed site outside Top 60 is appended"
-        );
-    }
-
-    #[test]
-    fn parse_excludes_official_and_unmanaged_non_top_rows() {
-        let page = parse_page(
-            LeaderboardKind::Claude,
-            include_str!("fixtures/veridrop-claude.html"),
-            &[],
-        )
-        .expect("parse fixture");
-
-        assert_eq!(
-            page.iter()
-                .map(|item| item.veridrop_host.as_str())
-                .collect::<Vec<_>>(),
-            vec!["bestapi.store"]
-        );
-    }
-
-    #[test]
-    fn parse_supports_all_public_leaderboard_kinds() {
-        let cases = [
-            (
-                LeaderboardKind::Overall,
-                include_str!("fixtures/veridrop-overall.html"),
-                "bestapi.store",
-                95,
-            ),
-            (
-                LeaderboardKind::OpenAi,
-                include_str!("fixtures/veridrop-openai.html"),
-                "bestapi.store",
-                95,
-            ),
-            (
-                LeaderboardKind::Gemini,
-                include_str!("fixtures/veridrop-gemini.html"),
-                "gemini.example",
-                98,
-            ),
-        ];
-
-        for (kind, fixture, host, score) in cases {
-            let items = parse_page(kind, fixture, &[]).expect("parse fixture");
-            assert_eq!(items.len(), 1, "{kind:?}");
-            assert_eq!(items[0].veridrop_host, host, "{kind:?}");
-            assert_eq!(items[0].score, score, "{kind:?}");
-        }
-    }
-
-    #[tokio::test]
-    async fn fetches_a_managed_detail_when_the_selected_leaderboard_omits_it() {
-        let app = Router::new()
-            .route(
-                "/leaderboard/gemini",
-                get(|| async { Html(include_str!("fixtures/veridrop-gemini.html")) }),
-            )
-            .route(
-                "/leaderboard/wawapii.com",
-                get(|| async { Html(include_str!("fixtures/veridrop-detail-wawapii.html")) }),
-            );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind local VeriDrop fixture server");
-        let address = listener.local_addr().expect("fixture server address");
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app)
-                .await
-                .expect("serve VeriDrop fixtures");
-        });
-        let client = reqwest::Client::builder()
-            .timeout(Duration::from_secs(2))
-            .build()
-            .expect("fixture client");
-
-        let items = fetch_live_source_with(
-            &client,
-            &format!("http://{address}"),
-            LeaderboardKind::Gemini,
-            &["wawapii.com".into()],
-        )
-        .await
-        .expect("fetch leaderboard with managed details");
-        server.abort();
-
-        let managed = items
-            .iter()
-            .find(|item| item.veridrop_host == "wawapii.com")
-            .expect("managed site with a Gemini score is appended");
-        assert_eq!(managed.rank, None, "detail pages do not own a rank");
-        assert_eq!(managed.score, 100);
-        assert_eq!(managed.samples, 1);
-        assert_eq!(managed.latest_date, "2026-08-13");
-        assert_eq!(managed.protocol_scores.len(), 3);
-        assert_eq!(managed.issues, vec!["token_usage"]);
-    }
-
-    #[tokio::test]
-    async fn managed_details_keep_the_configured_order_when_responses_finish_out_of_order() {
-        let app = Router::new()
-            .route(
-                "/leaderboard/gemini",
-                get(|| async { Html(include_str!("fixtures/veridrop-gemini.html")) }),
-            )
-            .route(
-                "/leaderboard/slow.example",
-                get(|| async {
-                    tokio::time::sleep(Duration::from_millis(50)).await;
-                    Html(include_str!("fixtures/veridrop-detail-wawapii.html"))
-                }),
-            )
-            .route(
-                "/leaderboard/fast.example",
-                get(|| async { Html(include_str!("fixtures/veridrop-detail-wawapii.html")) }),
-            );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind local VeriDrop fixture server");
-        let address = listener.local_addr().expect("fixture server address");
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app)
-                .await
-                .expect("serve VeriDrop fixtures");
-        });
-        let client = reqwest::Client::builder()
-            .timeout(Duration::from_secs(2))
-            .build()
-            .expect("fixture client");
-
-        let items = fetch_live_source_with(
-            &client,
-            &format!("http://{address}"),
-            LeaderboardKind::Gemini,
-            &["slow.example".into(), "fast.example".into()],
-        )
-        .await
-        .expect("fetch ordered managed details");
-        server.abort();
-
-        assert_eq!(
-            items
-                .iter()
-                .filter(|item| item.rank.is_none())
-                .map(|item| item.veridrop_host.as_str())
-                .collect::<Vec<_>>(),
-            vec!["slow.example", "fast.example"]
-        );
-    }
-
-    #[tokio::test]
-    async fn blocked_site_identity_does_not_fetch_its_veridrop_alias() {
-        let app = Router::new()
-            .route(
-                "/leaderboard/gemini",
-                get(|| async { Html(include_str!("fixtures/veridrop-gemini.html")) }),
-            )
-            .route(
-                "/leaderboard/api.790053500.com",
-                get(|| async { axum::http::StatusCode::INTERNAL_SERVER_ERROR }),
-            );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind local VeriDrop fixture server");
-        let address = listener.local_addr().expect("fixture server address");
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app)
-                .await
-                .expect("serve VeriDrop fixtures");
-        });
-        let client = reqwest::Client::builder()
-            .timeout(Duration::from_secs(2))
-            .build()
-            .expect("fixture client");
-        let mut config = config_with_directory();
-        config
-            .aff_codes
-            .insert("790053500.com".into(), "invite".into());
-        config.relay_directory.blocked_hosts = vec!["790053500.com".into()];
-
-        let items = fetch_live_source_with(
-            &client,
-            &format!("http://{address}"),
-            LeaderboardKind::Gemini,
-            &managed_veridrop_hosts(&config),
-        )
-        .await
-        .expect("blocked alias must not make the leaderboard fail");
-        server.abort();
-
-        assert_eq!(
-            items
-                .iter()
-                .map(|item| item.veridrop_host.as_str())
-                .collect::<Vec<_>>(),
-            vec!["gemini.example"]
-        );
-    }
-
-    #[test]
-    fn detail_samples_do_not_include_digits_from_the_host_name() {
-        let html = include_str!("fixtures/veridrop-detail-wawapii.html")
-            .replace("wawapii.com", "790053500.com");
-
-        let item = parse_detail_page(LeaderboardKind::Overall, "790053500.com", &html)
-            .expect("parse detail fixture")
-            .expect("overall score is present");
-
-        assert_eq!(item.samples, 276);
-    }
-
-    #[test]
-    fn leaderboard_kind_uses_frontend_tab_values() {
-        let cases = [
-            (LeaderboardKind::Overall, "\"overall\""),
-            (LeaderboardKind::Claude, "\"claude\""),
-            (LeaderboardKind::OpenAi, "\"openai\""),
-            (LeaderboardKind::Gemini, "\"gemini\""),
-        ];
-
-        for (kind, json) in cases {
-            assert_eq!(serde_json::to_string(&kind).unwrap(), json);
-            assert_eq!(serde_json::from_str::<LeaderboardKind>(json).unwrap(), kind);
-        }
-    }
-
-    #[test]
-    fn policy_maps_alias_entry_and_display_name_and_blocks_incompatible_hosts() {
-        let parsed = vec![
-            ParsedLeaderboardItem {
-                veridrop_host: "api.790053500.com".into(),
-                rank: Some(72),
-                score: 96,
-                samples: 9,
-                latest_date: "2026-08-11".into(),
-                detail_url: "https://veridrop.org/leaderboard/api.790053500.com".into(),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-            },
-            ParsedLeaderboardItem {
-                veridrop_host: "blocked.example".into(),
-                rank: Some(1),
-                score: 100,
-                samples: 99,
-                latest_date: "2026-08-13".into(),
-                detail_url: "https://veridrop.org/leaderboard/blocked.example".into(),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-            },
-        ];
-
-        let items = apply_policy(parsed, &config_with_directory(), LeaderboardKind::Overall);
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].site_host, "790053500.com");
-        assert_eq!(items[0].veridrop_host, "api.790053500.com");
-        assert_eq!(items[0].display_name, "鑫旺");
-        assert_eq!(items[0].entry_url, "https://790053500.com/keys");
-        assert!(items[0].auto_add);
-    }
-
-    #[test]
-    fn ordinary_leaderboard_sites_are_not_displayed() {
-        // 白名单语义（2026-08-15）：未受管的普通榜单站点没有过收录验证
-        // （探针可通 + 一键登录可用），**不进广场** —— 旧的「展示但 auto_add=false、
-        // 让用户去浏览器手动注册」路径已删。
-        let parsed = vec![ParsedLeaderboardItem {
-            veridrop_host: "ordinary.example".into(),
-            rank: Some(1),
-            score: 98,
-            samples: 12,
-            latest_date: "2026-08-14".into(),
-            detail_url: "https://veridrop.org/leaderboard/ordinary.example".into(),
-            protocol_scores: vec![],
-            claude_signature_rate: None,
-            scenarios: vec![],
-            issues: vec![],
-        }];
-
-        let items = apply_policy(parsed, &RemoteConfig::default(), LeaderboardKind::Overall);
-
-        assert!(items.is_empty(), "未受管站点不该出现在广场");
-    }
-
-    #[test]
-    fn sponsor_sites_are_displayed_with_one_click_login() {
-        // sponsors 声明的站点即使没有 relay_directory 条目，也要进广场且可一键登录：
-        // entry_url 回落到 sponsor 的 site origin。
-        let config = RemoteConfig {
-            sponsors: vec![crate::relay::remote_config::Sponsor {
-                site_origin: "https://wawazz.xyz".into(),
-                display_name: "WAWA ZZ API".into(),
-                tagline: String::new(),
-            }],
-            ..RemoteConfig::default()
-        };
-        let row = |veridrop_host: &str| ParsedLeaderboardItem {
-            veridrop_host: veridrop_host.into(),
-            rank: Some(1),
-            score: 96,
-            samples: 9,
-            latest_date: "2026-08-14".into(),
-            detail_url: format!("https://veridrop.org/leaderboard/{veridrop_host}"),
-            protocol_scores: vec![],
-            claude_signature_rate: None,
-            scenarios: vec![],
-            issues: vec![],
-        };
-
-        let items = apply_policy(
-            vec![row("wawazz.xyz"), row("top.example")],
-            &config,
-            LeaderboardKind::Overall,
-        );
-
-        assert_eq!(items.len(), 1, "只有受管站点保留");
-        assert_eq!(items[0].site_host, "wawazz.xyz");
-        assert_eq!(items[0].entry_url, "https://wawazz.xyz");
-        assert!(items[0].auto_add, "受管站点一律可一键登录");
-    }
-
-    #[test]
-    fn policy_keeps_only_the_first_ranked_row_for_each_site_identity() {
-        let row = |veridrop_host: &str, rank: u32| ParsedLeaderboardItem {
-            veridrop_host: veridrop_host.into(),
-            rank: Some(rank),
-            score: 96,
-            samples: 9,
-            latest_date: "2026-08-11".into(),
-            detail_url: format!("https://veridrop.org/leaderboard/{veridrop_host}"),
-            protocol_scores: vec![],
-            claude_signature_rate: None,
-            scenarios: vec![],
-            issues: vec![],
-        };
-
-        let items = apply_policy(
-            vec![row("api.790053500.com", 12), row("790053500.com", 72)],
-            &config_with_directory(),
-            LeaderboardKind::Overall,
-        );
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].site_host, "790053500.com");
-        assert_eq!(items[0].rank, Some(12));
-    }
-
-    #[test]
-    fn managed_sites_missing_from_veridrop_still_get_a_row_on_the_overall_board() {
-        // 白名单是展示的充分条件（2026-09-06 维护者拍板）：受管站不在 veridrop
-        // 榜/详情页 ⇒ 只是缺观测数据，行必须在 —— 手动添加入口已删的当下，
-        // 广场行是用户接入受管站的唯一入口。
-        let items = apply_policy(vec![], &config_with_directory(), LeaderboardKind::Overall);
-
-        assert_eq!(items.len(), 1, "blocked 站除外，受管站都要有行");
-        let item = &items[0];
-        assert_eq!(item.site_host, "790053500.com");
-        assert_eq!(item.display_name, "鑫旺");
-        assert_eq!(item.entry_url, "https://790053500.com/keys");
-        assert_eq!(item.veridrop_host, "api.790053500.com");
-        assert!(item.auto_add);
-        assert_eq!(item.rank, None);
-        assert_eq!(item.score, None);
-        assert_eq!(item.samples, None);
-        assert_eq!(item.latest_date, None);
-        assert_eq!(item.detail_url, None);
-    }
-
-    #[test]
-    fn protocol_boards_do_not_synthesize_rows() {
-        // 协议分榜是「该协议有观测数据的站」的排序视图，合成行只属于总榜。
-        let items = apply_policy(vec![], &config_with_directory(), LeaderboardKind::Claude);
-
-        assert!(items.is_empty());
-    }
-
-    #[test]
-    fn one_site_domain_one_synthesized_row_even_when_whitelisted_twice() {
-        // sponsors 与 relay_directory 常各录一形（www. vs 裸域）：同一注册域是
-        // 一站一行。受管名单的集合序里裸域先到，目录条目的名字/入口优先。
-        let mut config = config_with_directory();
-        config.sponsors = vec![crate::relay::remote_config::Sponsor {
-            site_origin: "https://www.790053500.com".into(),
-            display_name: "sponsor 形态".into(),
-            tagline: String::new(),
-        }];
-
-        let items = apply_policy(vec![], &config, LeaderboardKind::Overall);
-
-        assert_eq!(items.len(), 1, "同一注册域不能合成两行");
-        assert_eq!(items[0].display_name, "鑫旺");
-    }
-
-    #[test]
-    fn aff_only_sites_synthesize_a_row_named_by_host() {
-        // 只在 aff_codes 里、没录 relay_directory 的站也是受管站：名字与入口
-        // 回落 host 本身（想更体面随时补目录条目）。
-        let config = RemoteConfig {
-            aff_codes: BTreeMap::from([("aijws.example".to_string(), "CODE".to_string())]),
-            ..RemoteConfig::default()
-        };
-
-        let items = apply_policy(vec![], &config, LeaderboardKind::Overall);
-
-        assert_eq!(items.len(), 1);
-        assert_eq!(items[0].site_host, "aijws.example");
-        assert_eq!(items[0].display_name, "aijws.example");
-        assert_eq!(items[0].entry_url, "https://aijws.example");
-    }
-
-    /// 用仓内真实远端配置（`remote-config/public/v1/config.json`，线上部署那份的
-    /// 源）做端到端实证：veridrop 完全没有数据时（parsed 为空），**每一个受管
-    /// 站点都在总榜有行、都能一键登录**。这是「白名单 = 展示充分条件」的总闸 ——
-    /// 配置侧键形漂移、身份归一化回退、合成逻辑被改坏，都会在这里红。
-    #[test]
-    fn every_site_in_the_shipped_config_gets_a_row_with_zero_veridrop_data() {
-        let path = std::path::Path::new(env!("CARGO_MANIFEST_DIR"))
-            .join("../remote-config/public/v1/config.json");
-        let config: RemoteConfig = serde_json::from_slice(
-            &std::fs::read(&path).unwrap_or_else(|error| panic!("读不到 {:?}: {error}", path)),
-        )
-        .expect("仓内 v1 配置必须能被客户端 schema 解析");
-
-        let items = apply_policy(vec![], &config, LeaderboardKind::Overall);
-
-        let mut covered_domains: BTreeSet<String> =
-            items.iter().map(|item| item.site_domain.clone()).collect();
-        // blocked_hosts 是维护者的显式 kill switch，不算「漏斗拦」。
-        let blocked: BTreeSet<String> = config
-            .relay_directory
-            .blocked_hosts
-            .iter()
-            .map(|host| crate::relay::identity::site_domain(host))
-            .collect();
-        let managed_domains: BTreeSet<String> = managed_site_hosts(&config)
-            .into_iter()
-            .map(|host| crate::relay::identity::site_domain(&host))
-            .filter(|domain| !blocked.contains(domain))
-            .collect();
-        for domain in &managed_domains {
-            assert!(
-                covered_domains.remove(domain),
-                "受管域 {domain} 在零 veridrop 数据下没有广场行"
-            );
-        }
-        assert!(
-            covered_domains.is_empty(),
-            "多出了非受管的行: {covered_domains:?}"
-        );
-        assert!(
-            items.iter().all(|item| item.auto_add),
-            "每一行都必须可一键登录"
-        );
-    }
-
-    #[test]
-    fn cache_is_fresh_until_exactly_six_hours() {
-        let synced_at = 1_786_680_000;
-
-        assert!(is_fresh_at(synced_at, synced_at + 6 * 60 * 60 - 1));
-        assert!(!is_fresh_at(synced_at, synced_at + 6 * 60 * 60));
-    }
-
-    #[test]
-    fn cache_freshness_includes_the_managed_host_set() {
-        let mut config = config_with_directory();
-        let cached = CachedLeaderboard {
-            schema_version: CACHE_SCHEMA_VERSION,
-            kind: LeaderboardKind::Claude,
-            items: vec![],
-            managed_hosts: managed_veridrop_hosts(&config),
-            synced_at: 1_786_680_000,
-        };
-
-        assert!(is_fresh_for(&cached, cached.synced_at + 60, &config));
-
-        config.relay_directory.sites.insert(
-            "new.example".into(),
-            crate::relay::remote_config::RelayDirectorySite {
-                veridrop_host: Some("probe.new.example".into()),
-                entry_url: None,
-                purchase_url: None,
-                usage_url: None,
-                display_name: None,
-            },
-        );
-
-        assert!(!is_fresh_for(&cached, cached.synced_at + 60, &config));
-    }
-
-    #[tokio::test]
-    async fn concurrent_refreshes_share_one_successful_attempt() {
-        use std::sync::{atomic::AtomicUsize, Arc};
-
-        let coordinator = RefreshCoordinator::default();
-        let calls = Arc::new(AtomicUsize::new(0));
-        let cached = Arc::new(std::sync::Mutex::new(None));
-        let run = || {
-            let calls = calls.clone();
-            let cached_for_read = cached.clone();
-            let cached_for_write = cached.clone();
-            coordinator.run(
-                RefreshMode::Force,
-                false,
-                move || cached_for_read.lock().expect("cached result").clone(),
-                move || async move {
-                    calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                    tokio::time::sleep(Duration::from_millis(20)).await;
-                    let leaderboard = RelayLeaderboard {
-                        kind: LeaderboardKind::Claude,
-                        items: vec![],
-                        synced_at: 1,
-                    };
-                    *cached_for_write.lock().expect("cached result") = Some(leaderboard.clone());
-                    Ok(leaderboard)
-                },
-            )
-        };
-
-        let (first, second) = tokio::join!(run(), run());
-        let first = first.expect("first refresh result");
-        let second = second.expect("second refresh result");
-
-        assert_eq!(calls.load(std::sync::atomic::Ordering::SeqCst), 1);
-        assert_ne!(first.updated, second.updated);
-    }
-
-    #[tokio::test]
-    async fn concurrent_refreshes_share_one_failed_attempt() {
-        use std::sync::{atomic::AtomicUsize, Arc};
-
-        let coordinator = RefreshCoordinator::default();
-        let calls = Arc::new(AtomicUsize::new(0));
-        let run = || {
-            let calls = calls.clone();
-            coordinator.run(
-                RefreshMode::IfStale,
-                false,
-                || None,
-                move || async move {
-                    calls.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-                    tokio::time::sleep(Duration::from_millis(20)).await;
-                    Err(AppError::Config("offline".into()))
-                },
-            )
-        };
-
-        let (first, second) = tokio::join!(run(), run());
-
-        assert!(first.is_err());
-        assert!(second.is_err());
-        assert_eq!(calls.load(std::sync::atomic::Ordering::SeqCst), 1);
-    }
-
-    #[tokio::test]
-    async fn policy_change_during_refresh_starts_a_new_attempt() {
-        use std::sync::{atomic::AtomicUsize, Arc};
-
-        let coordinator = Arc::new(RefreshCoordinator::default());
-        let active_policy = Arc::new(AtomicUsize::new(1));
-        let cached_policy = Arc::new(AtomicUsize::new(0));
-        let calls = Arc::new(AtomicUsize::new(0));
-        let first_started = Arc::new(tokio::sync::Notify::new());
-        let release_first = Arc::new(tokio::sync::Notify::new());
-
-        let first = {
-            let coordinator = coordinator.clone();
-            let active_policy = active_policy.clone();
-            let cached_policy = cached_policy.clone();
-            let calls = calls.clone();
-            let first_started = first_started.clone();
-            let release_first = release_first.clone();
-            tokio::spawn(async move {
-                coordinator
-                    .run(
-                        RefreshMode::IfStale,
-                        true,
-                        || None,
-                        move || async move {
-                            calls.fetch_add(1, Ordering::SeqCst);
-                            let policy = active_policy.load(Ordering::SeqCst);
-                            first_started.notify_one();
-                            release_first.notified().await;
-                            cached_policy.store(policy, Ordering::SeqCst);
-                            Ok(RelayLeaderboard {
-                                kind: LeaderboardKind::Claude,
-                                items: vec![],
-                                synced_at: policy as i64,
-                            })
-                        },
-                    )
-                    .await
-            })
-        };
-
-        first_started.notified().await;
-        active_policy.store(2, Ordering::SeqCst);
-
-        let second = {
-            let coordinator = coordinator.clone();
-            let active_policy = active_policy.clone();
-            let cached_policy = cached_policy.clone();
-            let active_policy_for_fresh = active_policy.clone();
-            let cached_policy_for_fresh = cached_policy.clone();
-            let calls = calls.clone();
-            tokio::spawn(async move {
-                coordinator
-                    .run(
-                        RefreshMode::IfStale,
-                        true,
-                        || {
-                            (cached_policy_for_fresh.load(Ordering::SeqCst)
-                                == active_policy_for_fresh.load(Ordering::SeqCst))
-                            .then(|| RelayLeaderboard {
-                                kind: LeaderboardKind::Claude,
-                                items: vec![],
-                                synced_at: cached_policy_for_fresh.load(Ordering::SeqCst) as i64,
-                            })
-                        },
-                        move || async move {
-                            calls.fetch_add(1, Ordering::SeqCst);
-                            let policy = active_policy.load(Ordering::SeqCst);
-                            cached_policy.store(policy, Ordering::SeqCst);
-                            Ok(RelayLeaderboard {
-                                kind: LeaderboardKind::Claude,
-                                items: vec![],
-                                synced_at: policy as i64,
-                            })
-                        },
-                    )
-                    .await
-            })
-        };
-
-        release_first.notify_one();
-        first.await.expect("first task").expect("first refresh");
-        second.await.expect("second task").expect("second refresh");
-
-        assert_eq!(calls.load(Ordering::SeqCst), 2);
-        assert_eq!(cached_policy.load(Ordering::SeqCst), 2);
-    }
-
-    #[tokio::test]
-    async fn missing_cache_bypasses_recent_automatic_failure() {
-        use std::sync::{atomic::AtomicUsize, Arc};
-
-        let coordinator = RefreshCoordinator::default();
-        let calls = Arc::new(AtomicUsize::new(0));
-
-        let first_calls = calls.clone();
-        let first = coordinator
-            .run(
-                RefreshMode::IfStale,
-                false,
-                || None,
-                move || async move {
-                    first_calls.fetch_add(1, Ordering::SeqCst);
-                    Err(AppError::Config("offline".into()))
-                },
-            )
-            .await;
-        assert!(first.is_err());
-
-        let second_calls = calls.clone();
-        let second = coordinator
-            .run(
-                RefreshMode::IfStale,
-                false,
-                || None,
-                move || async move {
-                    second_calls.fetch_add(1, Ordering::SeqCst);
-                    Ok(RelayLeaderboard {
-                        kind: LeaderboardKind::Claude,
-                        items: vec![],
-                        synced_at: 2,
-                    })
-                },
-            )
-            .await
-            .expect("foreground read retries without a cache");
-
-        assert!(second.updated);
-        assert_eq!(calls.load(Ordering::SeqCst), 2);
-    }
-
-    #[test]
-    fn relay_leaderboard_serialization_has_no_from_cache_mirror() {
-        let leaderboard = RelayLeaderboard {
-            kind: LeaderboardKind::Claude,
-            items: vec![],
-            synced_at: 1_786_680_000,
-        };
-
-        let value = serde_json::to_value(leaderboard).expect("serialize leaderboard");
-
-        assert!(value.get("fromCache").is_none());
-        assert_eq!(value["syncedAt"], 1_786_680_000);
-    }
-
-    #[tokio::test]
-    #[serial]
-    async fn failed_refresh_preserves_previous_cache_bytes_and_timestamp() {
-        let temp = tempfile::tempdir().expect("create isolated home");
-        let _home = TestHomeGuard::set(temp.path());
-        let previous = CachedLeaderboard {
-            schema_version: CACHE_SCHEMA_VERSION,
-            kind: LeaderboardKind::Claude,
-            items: vec![ParsedLeaderboardItem {
-                veridrop_host: "bestapi.store".into(),
-                rank: Some(1),
-                score: 99,
-                samples: 20,
-                latest_date: "2026-08-12".into(),
-                detail_url: "https://veridrop.org/leaderboard/bestapi.store".into(),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-            }],
-            managed_hosts: vec![],
-            synced_at: 1_786_680_000,
-        };
-        write_cache(&previous).expect("write previous cache");
-        let path = cache_path(LeaderboardKind::Claude);
-        let previous_bytes = std::fs::read(&path).expect("read previous cache");
-
-        let app = Router::new().route(
-            "/leaderboard/claude",
-            get(|| async { axum::http::StatusCode::INTERNAL_SERVER_ERROR }),
-        );
-        let listener = tokio::net::TcpListener::bind("127.0.0.1:0")
-            .await
-            .expect("bind local VeriDrop fixture server");
-        let address = listener.local_addr().expect("fixture server address");
-        let server = tokio::spawn(async move {
-            axum::serve(listener, app)
-                .await
-                .expect("serve VeriDrop fixtures");
-        });
-        let client = reqwest::Client::builder()
-            .timeout(Duration::from_secs(2))
-            .build()
-            .expect("fixture client");
-
-        let result = refresh_with(
-            &client,
-            &format!("http://{address}"),
-            LeaderboardKind::Claude,
-            &RemoteConfig::default(),
-        )
-        .await;
-        server.abort();
-
-        assert!(result.is_err());
-        assert_eq!(
-            std::fs::read(&path).expect("read preserved cache"),
-            previous_bytes
-        );
-        assert_eq!(
-            read_cache(LeaderboardKind::Claude)
-                .expect("preserved cache")
-                .synced_at,
-            previous.synced_at
-        );
-    }
-
-    #[test]
-    fn managed_hosts_use_the_policy_owned_veridrop_identity_once() {
-        let mut config = config_with_directory();
-        config
-            .aff_codes
-            .insert("790053500.com".into(), "invite".into());
-
-        assert_eq!(managed_veridrop_hosts(&config), vec!["api.790053500.com"]);
-    }
-
-    #[test]
-    fn cached_items_are_filtered_by_the_current_signed_policy() {
-        let cached = CachedLeaderboard {
-            schema_version: CACHE_SCHEMA_VERSION,
-            kind: LeaderboardKind::Claude,
-            items: vec![ParsedLeaderboardItem {
-                veridrop_host: "blocked.example".into(),
-                rank: Some(1),
-                score: 100,
-                samples: 99,
-                latest_date: "2026-08-13".into(),
-                detail_url: "https://veridrop.org/leaderboard/blocked.example".into(),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-            }],
-            managed_hosts: vec![],
-            synced_at: 1,
-        };
-
-        let filtered = apply_policy_to_cached(cached, &config_with_directory());
-
-        assert!(filtered.items.is_empty());
-    }
-
-    #[test]
-    fn cached_items_use_the_current_auto_add_policy() {
-        let cached = CachedLeaderboard {
-            schema_version: CACHE_SCHEMA_VERSION,
-            kind: LeaderboardKind::Claude,
-            items: vec![ParsedLeaderboardItem {
-                veridrop_host: "api.790053500.com".into(),
-                rank: Some(72),
-                score: 96,
-                samples: 9,
-                latest_date: "2026-08-11".into(),
-                detail_url: "https://veridrop.org/leaderboard/api.790053500.com".into(),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-            }],
-            managed_hosts: vec![],
-            synced_at: 1,
-        };
-
-        let managed = apply_policy_to_cached(cached.clone(), &config_with_directory());
-        let unmanaged = apply_policy_to_cached(cached, &RemoteConfig::default());
-
-        assert_eq!(managed.items.len(), 1);
-        assert!(managed.items[0].auto_add);
-        assert!(
-            unmanaged.items.is_empty(),
-            "配置撤掉受管身份后，缓存里的旧站点也不该再展示"
-        );
-    }
-
-    #[test]
-    fn cached_veridrop_facts_can_be_restored_after_a_site_is_unblocked() {
-        let source = ParsedLeaderboardItem {
-            veridrop_host: "blocked.example".into(),
-            rank: Some(1),
-            score: 100,
-            samples: 99,
-            latest_date: "2026-08-13".into(),
-            detail_url: "https://veridrop.org/leaderboard/blocked.example".into(),
-            protocol_scores: vec![],
-            claude_signature_rate: None,
-            scenarios: vec![],
-            issues: vec![],
-        };
-        let cached: CachedLeaderboard = serde_json::from_slice(
-            &serde_json::to_vec(&CachedLeaderboard {
-                schema_version: CACHE_SCHEMA_VERSION,
-                kind: LeaderboardKind::Claude,
-                items: vec![source],
-                managed_hosts: vec![],
-                synced_at: 1,
-            })
-            .expect("serialize current cache shape"),
-        )
-        .expect("read current cache shape");
-
-        let restored = apply_policy_to_cached(
-            cached,
-            &RemoteConfig {
-                // 解除 block 后该站要重新可见，前提是它仍在受管名单里
-                // （sponsors / aff / promo / relay_directory 任一来源）。
-                sponsors: vec![crate::relay::remote_config::Sponsor {
-                    site_origin: "https://blocked.example".into(),
-                    display_name: "Blocked".into(),
-                    tagline: String::new(),
-                }],
-                ..RemoteConfig::default()
-            },
-        );
-
-        assert_eq!(restored.items.len(), 1);
-        assert_eq!(restored.items[0].site_host, "blocked.example");
-    }
-
-    #[test]
-    fn legacy_policy_filtered_cache_is_not_accepted_as_source_facts() {
-        let legacy = serde_json::json!({
-            "kind": "claude",
-            "items": [{
-                "siteHost": "bestapi.store",
-                "veridropHost": "bestapi.store",
-                "displayName": "BestAPI",
-                "rank": 1,
-                "score": 99,
-                "samples": 20,
-                "latestDate": "2026-08-12",
-                "detailUrl": "https://veridrop.org/leaderboard/bestapi.store",
-                "protocolScores": [],
-                "claudeSignatureRate": null,
-                "scenarios": [],
-                "issues": [],
-                "entryUrl": "https://bestapi.store"
-            }],
-            "syncedAt": 1,
-            "fromCache": true
-        });
-
-        assert!(serde_json::from_value::<CachedLeaderboard>(legacy).is_err());
-    }
-
-    #[tokio::test]
-    #[ignore = "需要外网；手动跑 --ignored"]
-    async fn live_public_pages_match_the_parser_contract() {
-        for kind in [
-            LeaderboardKind::Overall,
-            LeaderboardKind::Claude,
-            LeaderboardKind::OpenAi,
-            LeaderboardKind::Gemini,
-        ] {
-            let items = fetch_live_source(kind, &["api.790053500.com".into()])
-                .await
-                .expect("parse public leaderboard and managed details");
-            assert!(!items.is_empty(), "{kind:?} should contain visible rows");
-            assert!(
-                items
-                    .iter()
-                    .any(|item| item.veridrop_host == "bestapi.store"),
-                "{kind:?} should include the managed BestAPI site"
-            );
-            assert!(items.iter().all(|item| item.score <= 100));
-        }
-
-        let gemini = fetch_live_source(LeaderboardKind::Gemini, &["wawapii.com".into()])
-            .await
-            .expect("fetch Gemini leaderboard and managed detail");
-        let wawapi = gemini
-            .iter()
-            .find(|item| item.veridrop_host == "wawapii.com")
-            .expect("WawAPI should be completed from its public detail page");
-        assert_eq!(wawapi.rank, None);
-        assert!(wawapi
-            .protocol_scores
-            .iter()
-            .any(|protocol| protocol.protocol == "Gemini"));
-    }
-
-    /// 曝光闸：连续多轮「协议认不出」的站从广场摘掉；网络失败 / 没探过的站不摘。
-    /// 钉的是 `filter_probe_gated` 与 `SiteProbeStore::should_expose` 的合谋行为。
-    #[test]
-    fn probe_gate_hides_sites_with_consecutive_panel_misses() {
-        use crate::relay::site_probe::{
-            ProbeOutcome, ProbeVerdict, SiteProbeStore, HIDE_AFTER_CONSECUTIVE_PANEL_MISSES,
-        };
-
-        fn outcome(verdict: ProbeVerdict) -> ProbeOutcome {
-            ProbeOutcome {
-                verdict,
-                backend: None,
-                detail: "test".into(),
-                probed_at: 1,
-            }
-        }
-
-        let mut store = SiteProbeStore::default();
-        for _ in 0..HIDE_AFTER_CONSECUTIVE_PANEL_MISSES {
-            store.record("koozhan.example", &outcome(ProbeVerdict::UnrecognizedPanel));
-        }
-        store.record("cf-blocked.example", &outcome(ProbeVerdict::NetworkBlocked));
-
-        fn gated_item(site_host: &str) -> RelayDirectoryItem {
-            RelayDirectoryItem {
-                site_host: site_host.into(),
-                site_domain: crate::relay::identity::site_domain(site_host),
-                veridrop_host: site_host.into(),
-                display_name: site_host.into(),
-                rank: None,
-                score: Some(90),
-                samples: Some(30),
-                latest_date: Some("2026-08-16".into()),
-                detail_url: Some(format!("https://veridrop.org/site/{site_host}")),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-                entry_url: format!("https://{site_host}"),
-                auto_add: true,
-                transit: None,
-            }
-        }
-
-        let items = vec![
-            gated_item("koozhan.example"),
-            gated_item("cf-blocked.example"),
-            gated_item("fresh.example"),
-        ];
-        let remaining = filter_probe_gated(items, &store)
-            .into_iter()
-            .map(|item| item.site_host)
-            .collect::<Vec<_>>();
-
-        assert!(!remaining.contains(&"koozhan.example".to_owned()));
-        assert!(remaining.contains(&"cf-blocked.example".to_owned()));
-        assert!(remaining.contains(&"fresh.example".to_owned()));
-    }
-
-    /// 三份清单的唯一源：sponsors / aff / promo / directory **全部**要进这份名单，
-    /// 统一归一（www、大小写）并去 blocked。**会红的改法**：把任何一类从并集里
-    /// 拿掉 —— 那正是 wawapi.top 式的洞（aff 名单进了广场，探针/导入闸却漏了它）。
-    #[test]
-    fn managed_site_hosts_spans_sponsors_aff_promo_and_directory() {
-        let config = RemoteConfig {
-            relay_directory: RelayDirectoryPolicy {
-                blocked_hosts: vec!["blocked.example".into()],
-                sites: BTreeMap::from([(
-                    "bestapi.store".into(),
-                    crate::relay::remote_config::RelayDirectorySite::default(),
-                )]),
-            },
-            sponsors: vec![crate::relay::remote_config::Sponsor {
-                site_origin: "https://www.WawAPII.com".into(),
-                display_name: "WawAPI".into(),
-                tagline: String::new(),
-            }],
-            aff_codes: BTreeMap::from([("wawapi.top".to_string(), "AFF".into())]),
-            promo_codes: BTreeMap::from([("promo.example".to_string(), "PROMO".into())]),
-            ..RemoteConfig::default()
-        };
-
-        let hosts = managed_site_hosts(&config);
-        for expected in [
-            "bestapi.store",
-            "wawapii.com",
-            "wawapi.top",
-            "promo.example",
-        ] {
-            assert!(
-                hosts.contains(&expected.to_string()),
-                "受管全集缺 {expected}：{hosts:?}"
-            );
-        }
-        assert!(!hosts.contains(&"blocked.example".to_string()));
-    }
-
-    /// 漏斗摘掉中间名次的站后，veridrop 原始名次对用户是乱序（#2 直接跳 #7）。
-    /// 重排钉三件事：评分倒序、名次连续（1..N）、并列确定性
-    /// （同分先看 veridrop 原名次，无名次的回填站垫后，仍并列按 host 钉死）。
-    /// transit 摘要按归一 host 精确 join：有摘要的站带上、没有的保持
-    /// `None`（徽章不渲染），摘要缓存为空时整体短路（新装首启的常态）。
-    #[test]
-    #[serial]
-    fn decorate_transit_joins_by_host_and_short_circuits_when_empty() {
-        fn row(site_host: &str) -> RelayDirectoryItem {
-            RelayDirectoryItem {
-                site_host: site_host.into(),
-                site_domain: crate::relay::identity::site_domain(site_host),
-                veridrop_host: site_host.into(),
-                display_name: site_host.into(),
-                rank: None,
-                score: Some(90),
-                samples: Some(30),
-                latest_date: Some("2026-08-21".into()),
-                detail_url: Some(format!("https://veridrop.org/site/{site_host}")),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-                entry_url: format!("https://{site_host}"),
-                auto_add: true,
-                transit: None,
-            }
-        }
-
-        // 空缓存短路：一条都不该被碰（短路分支连 entries 都不查）。
-        let untouched = decorate_transit(vec![row("a.example")]);
-        assert!(untouched[0].transit.is_none());
-
-        // 写一份只含 a.example 的摘要缓存，join 后只有它带徽章数据。
-        // guard 必须活到断言之后：提前 Drop 会还原 home、让 decorate 读到
-        // 真实用户目录里的缓存。
-        let _guard = crate::relay::transit::tests::transit_cache_guard("decorate");
-        crate::relay::transit::tests::write_transit_cache_entry(
-            "a.example",
-            crate::relay::transit::tests::badge_summary(Some(0.06), Some(95.0)),
-        );
-        let decorated = decorate_transit(vec![row("a.example"), row("b.example")]);
-        let summary = decorated[0].transit.clone().expect("a.example 应带上摘要");
-        assert_eq!(summary.min_multiplier, Some(0.06));
-        assert!(decorated[1].transit.is_none(), "没有摘要的站保持 None");
-    }
-
-    #[test]
-    fn renumber_ranks_by_score_closes_funnel_gaps() {
-        fn row(site_host: &str, score: Option<u8>, rank: Option<u32>) -> RelayDirectoryItem {
-            RelayDirectoryItem {
-                site_host: site_host.into(),
-                site_domain: crate::relay::identity::site_domain(site_host),
-                veridrop_host: site_host.into(),
-                display_name: site_host.into(),
-                rank,
-                score,
-                samples: Some(30),
-                latest_date: Some("2026-08-16".into()),
-                detail_url: Some(format!("https://veridrop.org/site/{site_host}")),
-                protocol_scores: vec![],
-                claude_signature_rate: None,
-                scenarios: vec![],
-                issues: vec![],
-                entry_url: format!("https://{site_host}"),
-                auto_add: true,
-                transit: None,
-            }
-        }
-
-        // 名次带洞（2/7/None/1），评分乱序给出；无评分的是合成行，垫在一切
-        // 有评分者之后。
-        let items = vec![
-            row("gap.example", Some(88), Some(7)),
-            row("tie-a.example", Some(95), Some(2)),
-            row("backfilled.example", Some(92), None),
-            row("tie-b.example", Some(95), Some(1)),
-            row("synthesized.example", None, None),
-        ];
-
-        let renumbered = renumber_ranks_by_score(items);
-        let order = renumbered
-            .iter()
-            .map(|item| (item.site_host.as_str(), item.rank))
-            .collect::<Vec<_>>();
-
-        assert_eq!(
-            order,
-            vec![
-                ("tie-b.example", Some(1)),       // 95 分、veridrop 原名次 1 → 并列在前
-                ("tie-a.example", Some(2)),       // 95 分、原名次 2
-                ("backfilled.example", Some(3)),  // 92 分、无原名次也不丢位置
-                ("gap.example", Some(4)),         // 88 分
-                ("synthesized.example", Some(5)), // 无观测数据，垫底但仍有行有名次
-            ],
-            "评分倒排 + 连续名次，洞被补上"
-        );
-    }
-}
