+++ src-tauri/src/relay/mod.rs
@@ -63,11 +63,11 @@ pub mod browser_bridge;
 pub mod cc_switch_import;
 pub mod chatgpt_app;
 pub mod creds;
+pub mod directory;
 pub mod discovery;
 pub mod identity;
 pub mod imagegen;
 pub mod imagegen_mcp;
-pub mod leaderboard;
 pub mod login;
 pub mod managed;
 pub mod newapi;
